// frontend/src/lib/api/model/levelEnum.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Explorito - Application Éducative
 * OpenAPI spec version: 1.0.0
 */

/**
 * Niveaux scolaires français, de la maternelle au CM2.
 */
export type LevelEnum = (typeof LevelEnum)[keyof typeof LevelEnum];

export const LevelEnum = {
  ps: "ps",
  ms: "ms",
  gs: "gs",
  cp: "cp",
  ce1: "ce1",
  ce2: "ce2",
  cm1: "cm1",
  cm2: "cm2",
} as const;